+++ src/components/input/SuggestInput/index.tsx
@@ -16,7 +16,12 @@ const Dropdown = styled(DropComponent)`
   padding: 8px 0;
 `;
 
-const SearchIcon = styled(SearchOutlineSVG)`
+function AnyIcon({ icon, ...props }: { icon: React.FunctionComponent<any>; onClick?: () => void; tabIndex?: number }) {
+  const Icon = icon;
+  return <Icon {...props} />;
+}
+
+const SearchIcon = styled(AnyIcon)`
   & *[fill^='#'] {
     fill: ${(props) => props.theme.color['Neutral/Neutral 50']};
   }
@@ -99,10 +104,18 @@ export interface SuggestInputProps extends Omit<TextInputProps, 'value'> {
 
   /** Задает максимальную высоту контейнера с опциями */
   dropMaxHeight?: string | number;
+
+  /**
+   * Компонент для отображения альтернативной иконки
+   */
+  icon?: React.FunctionComponent;
 }
 
 export const SuggestInput = React.forwardRef<HTMLInputElement, SuggestInputProps>(
-  ({ isLoading, onOptionSelect, alignDropdown, dropMaxHeight = '300px', ...props }, ref) => {
+  (
+    { isLoading, onOptionSelect, alignDropdown, dropMaxHeight = '300px', icons, icon = SearchOutlineSVG, ...props },
+    ref,
+  ) => {
     const isControlledComponentValue = undefined !== props.value;
     const { options, portalTargetRef } = props;
 
@@ -171,7 +184,14 @@ export const SuggestInput = React.forwardRef<HTMLInputElement, SuggestInputProps
 
     const loadingMessage = isLoading ? <SearchingPanel>Поиск совпадений</SearchingPanel> : 'Нет совпадений';
 
-    const delayBlurEffect = React.useCallback(() => window.setTimeout(() => setIsFocused(false), 200), [setIsFocused]);
+    const [blurTrigger, triggerDelayedBlur] = React.useState<undefined | unknown>();
+
+    React.useEffect(() => {
+      if (blurTrigger) {
+        const timeoutID = setTimeout(() => setIsFocused(false), 200);
+        return () => clearTimeout(timeoutID);
+      }
+    }, [blurTrigger, setIsFocused]);
 
     React.useEffect(() => {
       function onInputChange(this: HTMLInputElement) {
@@ -187,11 +207,14 @@ export const SuggestInput = React.forwardRef<HTMLInputElement, SuggestInputProps
       }
     }, [props.onInput]);
 
+    const iconArray = React.Children.toArray(icons);
+    iconArray.push(<SearchIcon icon={icon} />);
+
     return (
       <TextInput
         {...props}
         ref={refSetter(ref, inputRef)}
-        icons={<SearchIcon />}
+        icons={!props.readOnly && iconArray}
         onKeyUp={(...p) => {
           props.onKeyUp?.(...p);
           handleKeyUp(...p);
@@ -206,7 +229,7 @@ export const SuggestInput = React.forwardRef<HTMLInputElement, SuggestInputProps
         }}
         onBlur={(...p) => {
           props.onBlur?.(...p);
-          delayBlurEffect();
+          triggerDelayedBlur({});
         }}
       >
         {options && isSuggestPanelOpen && (
@@ -226,7 +249,7 @@ export const SuggestInput = React.forwardRef<HTMLInputElement, SuggestInputProps
                   active={index === activeIndex}
                   searchText={props.value || inputRef.current?.value || ''}
                   onMouseEnter={() => setActiveIndex(index)}
-                  onClick={handleOptionClick}
+                  onMouseDown={handleOptionClick}
                   onKeyUp={handleKeyUp}
                 />
               ))